MAJOR bug fixes; add lat/lon subsetting functionality
Fixes issue where one of the calls to estimate.quantiles flipped
the desired quantiles to fit between normalization, bulk, and tail
fits, resuling in unusable output and strange normalization behavior.

Lat/lon subsetting functionality is added through a list element in
[defaults] - if defaults$lat.clip or defaults$lon.clip are populated,
[get.process.chunks] will only keep pixels within those limits.
Global idx ordering is kept so that missing gaps can be filled in
later (creates possible issue if pixels are loaded in a non-monotonic
fashion... merits further consideration - some of the filenaming
conventions should maybe be changed to subset global idx in the future
to avoid the lack of reprocessing of skipped pixels in the same
global location chunk).