Adds workaround for bug in ncdf4 resulting in geographic drift
If the header is written *after* the main data to a NetCDF file
using ncdf4, sometimes if the file is big enough, part of the data
will be shifted over by about 90 degrees latitude at some random
timestep. Still diagnosing the problem, but a workaround in which
the header is written first in [merge_locs] is added, which fixes
the issue.